+++ CHANGELOG.md
@@ -17,3 +17,12 @@
 ## 1.1.1
 * 添加 `Number` 扩展的支持。
 * `Convert` 添加转换为DateTime、Date两种时间类型的支持。
+
+## 1.1.2
+* 适配TypeScript 4.X 中关于Number扩展中访问this报错的问题
+
+## 1.1.3
+* 修复Dictionary中import不合理的地方
+
+## 1.1.4
+* 修正 `Dictionary` 中构造函数类型推论错误的情况